# Varint to varchar optimization