--- a/src/common/types/varint.cpp
+++ b/src/common/types/varint.cpp
@@ -1,5 +1,7 @@
 #include "duckdb/common/types/varint.hpp"
 #include "duckdb/common/exception/conversion_exception.hpp"
+#include "duckdb/common/numeric_utils.hpp"
+#include "duckdb/common/typedefs.hpp"
 #include <cmath>
 
 namespace duckdb {
@@ -159,12 +161,15 @@ void Varint::GetByteArray(vector<uint8_t> &byte_array, bool &is_negative, const 
 
 	// Determine if the number is negative
 	is_negative = (blob_ptr[0] & 0x80) == 0;
-	for (idx_t i = 3; i < blob.GetSize(); i++) {
-		if (is_negative) {
+	byte_array.reserve(blob.GetSize() - 3);
+	if (is_negative) {
+		for (idx_t i = 3; i < blob.GetSize(); i++) {
 			byte_array.push_back(static_cast<uint8_t>(~blob_ptr[i]));
-		} else {
+		} 
+	} else {
+		for (idx_t i = 3; i < blob.GetSize(); i++) {
 			byte_array.push_back(static_cast<uint8_t>(blob_ptr[i]));
-		}
+		} 	
 	}
 }
 
@@ -184,28 +189,55 @@ string Varint::FromByteArray(uint8_t *data, idx_t size, bool is_negative) {
 	return result;
 }
 
+// Following CPython and Knuth (TAOCP, Volume 2 (3rd edn), section 4.4, Method 1b). 
 string Varint::VarIntToVarchar(const string_t &blob) {
 	string decimal_string;
 	vector<uint8_t> byte_array;
 	bool is_negative;
 	GetByteArray(byte_array, is_negative, blob);
-	while (!byte_array.empty()) {
-		string quotient;
-		uint8_t remainder = 0;
-		for (uint8_t byte : byte_array) {
-			int new_value = remainder * 256 + byte;
-			quotient += DigitToChar(new_value / 10);
-			remainder = static_cast<uint8_t>(new_value % 10);
+	vector<digit_t> digits;
+	// Rounding byte_array to digit_bytes multiple size, so that we can process every digit_bytes bytes 
+	// at a time without if check in the for loop
+	idx_t digit_bytes = sizeof(digit_t);
+	idx_t digit_bits = digit_bytes * 8;
+	idx_t padding_size = (digit_bytes - byte_array.size() % digit_bytes) % digit_bytes;
+	byte_array.insert(byte_array.begin(), padding_size, 0);
+	for (idx_t i = 0; i < byte_array.size(); i += digit_bytes) {
+		digit_t hi = 0;
+		for (idx_t j = 0; j < digit_bytes; j++) {
+			hi |= UnsafeNumericCast<digit_t>(byte_array[i+j]<< (8*(digit_bytes-j-1)));
 		}
-		decimal_string += DigitToChar(remainder);
-		// Remove leading zeros from the quotient
-		byte_array.clear();
-		for (char digit : quotient) {
-			if (digit != '0' || !byte_array.empty()) {
-				byte_array.push_back(static_cast<uint8_t>(CharToDigit(digit)));
-			}
+
+		for (idx_t j = 0; j < digits.size(); j++) {
+			twodigit_t tmp = UnsafeNumericCast<twodigit_t>(digits[j]) << digit_bits | hi;
+			hi = UnsafeNumericCast<digit_t>(tmp / UnsafeNumericCast<twodigit_t>(DECIMAL_BASE));
+			digits[j] = UnsafeNumericCast<digit_t>(tmp-UnsafeNumericCast<twodigit_t>(DECIMAL_BASE*hi));
+		}
+
+		while(hi) {
+			digits.push_back(hi % UnsafeNumericCast<digit_t>(DECIMAL_BASE));
+			hi /= UnsafeNumericCast<digit_t>(DECIMAL_BASE);
 		}
 	}
+
+	if (digits.size() == 0) {
+		digits.push_back(0);
+	}
+
+	for (idx_t i = 0; i < digits.size()-1; i++) {
+		auto remain = digits[i];
+		for (idx_t j = 0; j < DECIMAL_SHIFT; j++) {
+			decimal_string += DigitToChar(UnsafeNumericCast<int>(remain % 10));
+			remain /= 10;
+		}
+	}
+
+	auto remain = digits.back();
+	do {
+		decimal_string += DigitToChar(UnsafeNumericCast<int>(remain % 10));
+		remain /= 10;
+	} while (remain != 0);
+
 	if (is_negative) {
 		decimal_string += '-';
 	}