Following #13015, use the algorithm mentioned in CPython to optimize conversion from varint to varchar. Knuth (TAOCP, Volume 2 (3rd edn), section 4.4, Method 1b)
### Benchmarks
```sql
-- load
CREATE TABLE t1 AS select '340282366920938463463374607431768211455'::varint as a from range(0,10000000) tbl(i);

-- run
create or replace table t2 as select a::varchar from t1
```

Benchmark is ran on my laptop which is Mac M2/16G.

|old|new|
|---|----|
|3.22s|0.72s|